requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
pkg-resources==0.0.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.1
